rfix: fix imports and move track_catalog into data directory
